requirements.txt
fastapi
uvicorn[standard]
crewai
crewai-tools
langchain-openai
python-dotenv
celery
redis
PyGithub
pydantic
pydantic-settings